requirements.txt
appdirs==1.4.3
boto3==1.13.1
botocore==1.16.1
CacheControl==0.12.6
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2.5
django-cors-middleware==1.4.0
django-filter==2.2.0
django-storages==1.9.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
docutils==0.15.2
elasticsearch==7.8.1
elasticsearch-dsl==7.2.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jmespath==0.9.5
lockfile==0.12.2
Markdown==3.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.3
pycrypto==2.6.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2019.2
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.6
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
